% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of variation}
\usage{
cv(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
Returns a numeric value
}
\description{
Calculates the coefficient of variation (CV) for a vector of numbers
}
\examples{
cv(c(23, 54, 34, 10, 5))


}
\author{
LTER metacommunities group

Eric R. Sokol \email{sokole@gmail.com}
}
