% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divpart.R
\name{divpart}
\alias{divpart}
\title{Numbers equivalents for a metacommunity time series.}
\usage{
divpart(data_long = NULL, data_wide = NULL,
  time_step_col_name = NULL, site_id_col_name = NULL,
  taxon_list = NULL, taxon_id_col_name = NULL,
  biomass_col_name = NULL, q_value = 0)
}
\arguments{
\item{data_long}{Data input for analysis (should be \code{NULL} if input is wide format data). Long format data frame (or tibble) with species counts (or biomass measurements) for multiple sites observed over multiple time periods}

\item{data_wide}{Data input for analysis (should be \code{NULL} if input is long format data). Wide format data frame (or tibble) where rows are observations and columns include \code{time_step_col_name}, \code{site_id_col_name}, taxa listed in \code{taxon_list}.}

\item{time_step_col_name}{Character string indicating time step column name. For use with both \code{data_long} and \code{data_wide} inputs.}

\item{site_id_col_name}{Character string indicating site id column name. For use with both \code{data_long} and \code{data_wide} inputs.}

\item{taxon_list}{Vector (character) of taxon IDs, used to indicate which columns in \code{data_wide} are species counts (or biomass measurements)}

\item{taxon_id_col_name}{Character string indicating column name in \code{data_long} that contains taxon IDs.}

\item{biomass_col_name}{Character string indicating column name in \code{data_long} that contains biomass measurements (or species counts).}

\item{q_value}{Order of the diversity measure. Default is q = 0.}
}
\value{
Returns a one-row data frame where each element names indicate the variability metric that is being returned.
}
\description{
A wrapper function for \code{vegetarian::d()} to calculate "numbers equivalents" alpha, beta, and gamma diversity indicies for each time step in a metacommunity data set. See Jost (2007) and documentation for \code{vegetarian::d()} for more information about the metrics.
}
\examples{
# examples
set.seed(123)
library(dplyr)
library(tidyr)

dat_example_long <- data.frame(
  my_sites = c(rep('x', 8), rep('y',8)),
  my_times = rep(1:4, each = 2),
  my_spp = rep(c('a','b'), 4),
  my_biomass = (exp(rnorm(16))*10) \%>\% round(0))

divpart(
  data_long = dat_example_long,
  site_id_col_name = 'my_sites',
  time_step_col_name = 'my_times',
  taxon_id_col_name = 'my_spp',
  biomass_col_name = 'my_biomass',
  q_value = 0)

divpart(
  data_long = dat_example_long,
  site_id_col_name = 'my_sites',
  time_step_col_name = 'my_times',
  taxon_id_col_name = 'my_spp',
  biomass_col_name = 'my_biomass',
  q_value = 1)

divpart(
  data_long = dat_example_long,
  site_id_col_name = 'my_sites',
  time_step_col_name = 'my_times',
  taxon_id_col_name = 'my_spp',
  biomass_col_name = 'my_biomass',
  q_value = 2)

# examples with wide data
dat_example_wide <- dat_example_long \%>\% spread(my_spp, my_biomass)

divpart(
  data_wide = dat_example_wide,
  site_id_col_name = 'my_sites',
  time_step_col_name = 'my_times',
  taxon_list = c('a','b'),
  biomass_col_name = 'my_biomass',
  q_value = 0)

}
\references{
Jost, L. 2007. Partitioning diversity into independent alpha and beta components. Ecology 88:2427–2439. https://doi.org/10.1890/06-1736.1.
}
\author{
Eric R. Sokol \email{sokole@gmail.com}

Thomas Lamy \email{thomas.lamy27@gmail.com}
}
